+++ src/packages/esProjectMaker.js
@@ -1,5 +1,10 @@
 module.exports = {
-  base: [
-    {'@mparulski/es-project-maker': {version: '0.14.2', type: '--save-dev'}},
-  ],
+  "base": [
+    {
+      "@mparulski/es-project-maker": {
+        "version": "0.15.0",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }