+++ src/packages/eslint.js
@@ -1,6 +1,16 @@
 module.exports = {
-  base: [
-    {'@mparulski/es-project-eslint': {version: '0.1.0', type: '--save-dev'}},
-    {eslint: {version: '7.32.0', type: '--save-dev'}},
-  ],
+  "base": [
+    {
+      "@mparulski/es-project-eslint": {
+        "version": "0.1.0",
+        "type": "--save-dev"
+      }
+    },
+    {
+      "eslint": {
+        "version": "7.32.0",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }