+++ src/packages/prettier.js
@@ -1,3 +1,10 @@
 module.exports = {
-  base: [{prettier: {version: '2.4.1', type: '--save-dev'}}],
+  "base": [
+    {
+      "prettier": {
+        "version": "2.4.1",
+        "type": "--save-dev"
+      }
+    }
+  ]
 }