+++ src/packages/react.js
@@ -1,6 +1,16 @@
 module.exports = {
-  base: [
-    {react: {version: '17.0.2', type: '--save'}},
-    {'react-dom': {version: '17.0.2', type: '--save'}},
-  ],
+  "base": [
+    {
+      "react": {
+        "version": "17.0.2",
+        "type": "--save"
+      }
+    },
+    {
+      "react-dom": {
+        "version": "17.0.2",
+        "type": "--save"
+      }
+    }
+  ]
 }