+++ src/scripts/utils/isComparedVersionIsNewerThanInstalled.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const semver = require('semver')
+
+function isComparedVersionIsNewerThanInstalled(comparedVersion, installedVersion) {
+    return semver.gt(comparedVersion, installedVersion)
+}
+
+module.exports = isComparedVersionIsNewerThanInstalled